+++ tests/expandandmerge.cpp
@@ -90,12 +90,12 @@ void testFloat()
         d[i] = double_v::Random();
     }
     float_v f(d);
-    for (int i = 0; i < float_v::Size; ++i) {
+    for (size_t i = 0; i < float_v::Size; ++i) {
         COMPARE(f[i], float(d[i / double_v::Size][i % double_v::Size])) << ", i = " << i;
     }
     f = float_v::Random();
     f.expand(d);
-    for (int i = 0; i < float_v::Size; ++i) {
+    for (size_t i = 0; i < float_v::Size; ++i) {
         COMPARE(f[i], float(d[i / double_v::Size][i % double_v::Size])) << ", i = " << i;
     }
 }
