+++ tests/math.cpp
@@ -196,7 +196,7 @@ template<typename V> void testLog()/*{{{*/
     Array<Reference<T> > reference = referenceData<T, Log>();
     for (size_t i = 0; i + V::Size - 1 < reference.size; i += V::Size) {
         V x, ref;
-        for (int j = 0; j < V::Size; ++j) {
+        for (size_t j = 0; j < V::Size; ++j) {
             x[j] = reference.data[i + j].x;
             ref[j] = reference.data[i + j].ref;
         }
@@ -227,7 +227,7 @@ template<typename V> void testLog2()/*{{{*/
     Array<Reference<T> > reference = referenceData<T, Log2>();
     for (size_t i = 0; i + V::Size - 1 < reference.size; i += V::Size) {
         V x, ref;
-        for (int j = 0; j < V::Size; ++j) {
+        for (size_t j = 0; j < V::Size; ++j) {
             x[j] = reference.data[i + j].x;
             ref[j] = reference.data[i + j].ref;
         }
@@ -250,7 +250,7 @@ template<typename V> void testLog10()/*{{{*/
     Array<Reference<T> > reference = referenceData<T, Log10>();
     for (size_t i = 0; i + V::Size - 1 < reference.size; i += V::Size) {
         V x, ref;
-        for (int j = 0; j < V::Size; ++j) {
+        for (size_t j = 0; j < V::Size; ++j) {
             x[j] = reference.data[i + j].x;
             ref[j] = reference.data[i + j].ref;
         }
@@ -270,7 +270,7 @@ template<typename Vec> void testMax()/*{{{*/
     typedef typename Vec::EntryType T;
     VectorMemoryHelper<Vec> mem(3);
     T *data = mem;
-    for (int i = 0; i < Vec::Size; ++i) {
+    for (size_t i = 0; i < Vec::Size; ++i) {
         data[i] = i;
         data[i + Vec::Size] = Vec::Size + 1 - i;
         data[i + 2 * Vec::Size] = std::max(data[i], data[i + Vec::Size]);
@@ -286,7 +286,7 @@ template<typename Vec> void testMax()/*{{{*/
 #define FillHelperMemory(code) \
     typename V::Memory data; \
     typename V::Memory reference; \
-    for (int ii = 0; ii < V::Size; ++ii) { \
+    for (size_t ii = 0; ii < V::Size; ++ii) { \
         const T i = static_cast<T>(ii); \
         data[ii] = i; \
         reference[ii] = code; \
@@ -320,7 +320,7 @@ template<typename V> void testSincos()/*{{{*/
     Array<SincosReference<T> > reference = sincosReference<T>();
     for (size_t i = 0; i + V::Size - 1 < reference.size; i += V::Size) {
         V x, sref, cref;
-        for (int j = 0; j < V::Size; ++j) {
+        for (size_t j = 0; j < V::Size; ++j) {
             x[j] = reference.data[i + j].x;
             sref[j] = reference.data[i + j].s;
             cref[j] = reference.data[i + j].c;
@@ -343,7 +343,7 @@ template<typename V> void testSin()/*{{{*/
     Array<SincosReference<T> > reference = sincosReference<T>();
     for (size_t i = 0; i + V::Size - 1 < reference.size; i += V::Size) {
         V x, sref;
-        for (int j = 0; j < V::Size; ++j) {
+        for (size_t j = 0; j < V::Size; ++j) {
             x[j] = reference.data[i + j].x;
             sref[j] = reference.data[i + j].s;
         }
@@ -360,7 +360,7 @@ template<typename V> void testCos()/*{{{*/
     Array<SincosReference<T> > reference = sincosReference<T>();
     for (size_t i = 0; i + V::Size - 1 < reference.size; i += V::Size) {
         V x, cref;
-        for (int j = 0; j < V::Size; ++j) {
+        for (size_t j = 0; j < V::Size; ++j) {
             x[j] = reference.data[i + j].x;
             cref[j] = reference.data[i + j].c;
         }
@@ -377,7 +377,7 @@ template<typename V> void testAsin()/*{{{*/
     Array<Reference<T> > reference = referenceData<T, Asin>();
     for (size_t i = 0; i + V::Size - 1 < reference.size; i += V::Size) {
         V x, ref;
-        for (int j = 0; j < V::Size; ++j) {
+        for (size_t j = 0; j < V::Size; ++j) {
             x[j] = reference.data[i + j].x;
             ref[j] = reference.data[i + j].ref;
         }
@@ -413,7 +413,7 @@ template<typename V> void testAtan()/*{{{*/
     Array<Reference<T> > reference = referenceData<T, Atan>();
     for (size_t i = 0; i + V::Size - 1 < reference.size; i += V::Size) {
         V x, ref;
-        for (int j = 0; j < V::Size; ++j) {
+        for (size_t j = 0; j < V::Size; ++j) {
             x[j] = reference.data[i + j].x;
             ref[j] = reference.data[i + j].ref;
         }
@@ -507,7 +507,7 @@ template<typename Vec> void testReciprocal()/*{{{*/
         const T scale = T(0.1);
         typename Vec::Memory data;
         typename Vec::Memory reference;
-        for (int ii = 0; ii < Vec::Size; ++ii) {
+        for (size_t ii = 0; ii < Vec::Size; ++ii) {
             const T i = static_cast<T>(ii);
             data[ii] = i;
             T tmp = (i + offset) * scale;
@@ -568,7 +568,7 @@ template<typename Vec> void testRound()/*{{{*/
     VectorMemoryHelper<Vec> mem2(Count);
     T *data = mem1;
     T *reference = mem2;
-    for (int i = 0; i < Count * Vec::Size; ++i) {
+    for (size_t i = 0; i < Count * Vec::Size; ++i) {
         data[i] = i * 0.25 - 2.0;
         reference[i] = std::floor(i * 0.25 - 2.0 + 0.5);
         if (i % 8 == 2) {
@@ -591,10 +591,10 @@ template<typename Vec> void testReduceMin()/*{{{*/
     const T one = 1;
     VectorMemoryHelper<Vec> mem(Vec::Size);
     T *data = mem;
-    for (int i = 0; i < Vec::Size * Vec::Size; ++i) {
+    for (size_t i = 0; i < Vec::Size * Vec::Size; ++i) {
         data[i] = i % (Vec::Size + 1) + one;
     }
-    for (int i = 0; i < Vec::Size; ++i, data += Vec::Size) {
+    for (size_t i = 0; i < Vec::Size; ++i, data += Vec::Size) {
         const Vec a(&data[0]);
         //std::cout << a << std::endl;
         COMPARE(a.min(), one);
@@ -607,10 +607,10 @@ template<typename Vec> void testReduceMax()/*{{{*/
     const T max = Vec::Size + 1;
     VectorMemoryHelper<Vec> mem(Vec::Size);
     T *data = mem;
-    for (int i = 0; i < Vec::Size * Vec::Size; ++i) {
+    for (size_t i = 0; i < Vec::Size * Vec::Size; ++i) {
         data[i] = (i + Vec::Size) % (Vec::Size + 1) + 1;
     }
-    for (int i = 0; i < Vec::Size; ++i, data += Vec::Size) {
+    for (size_t i = 0; i < Vec::Size; ++i, data += Vec::Size) {
         const Vec a(&data[0]);
         //std::cout << a << std::endl;
         COMPARE(a.max(), max);
@@ -624,16 +624,16 @@ template<typename Vec> void testReduceProduct()/*{{{*/
     };
     typedef typename Vec::EntryType T;
     int _product = 1;
-    for (int i = 1; i < Vec::Size; ++i) {
+    for (size_t i = 1; i < Vec::Size; ++i) {
         _product *= (i % Max) + 1;
     }
     const T product = _product;
     VectorMemoryHelper<Vec> mem(Vec::Size);
     T *data = mem;
-    for (int i = 0; i < Vec::Size * Vec::Size; ++i) {
+    for (size_t i = 0; i < Vec::Size * Vec::Size; ++i) {
         data[i] = ((i + (i / Vec::Size)) % Max) + 1;
     }
-    for (int i = 0; i < Vec::Size; ++i, data += Vec::Size) {
+    for (size_t i = 0; i < Vec::Size; ++i, data += Vec::Size) {
         const Vec a(&data[0]);
         //std::cout << a << std::endl;
         COMPARE(a.product(), product);
@@ -644,16 +644,16 @@ template<typename Vec> void testReduceSum()/*{{{*/
 {
     typedef typename Vec::EntryType T;
     int _sum = 1;
-    for (int i = 2; i <= Vec::Size; ++i) {
+    for (size_t i = 2; i <= Vec::Size; ++i) {
         _sum += i;
     }
     const T sum = _sum;
     VectorMemoryHelper<Vec> mem(Vec::Size);
     T *data = mem;
-    for (int i = 0; i < Vec::Size * Vec::Size; ++i) {
+    for (size_t i = 0; i < Vec::Size * Vec::Size; ++i) {
         data[i] = (i + i / Vec::Size) % Vec::Size + 1;
     }
-    for (int i = 0; i < Vec::Size; ++i, data += Vec::Size) {
+    for (size_t i = 0; i < Vec::Size; ++i, data += Vec::Size) {
         const Vec a(&data[0]);
         //std::cout << a << std::endl;
         COMPARE(a.sum(), sum);
@@ -803,7 +803,7 @@ template<typename V> void testUlpDiff()/*{{{*/
 
             VERIFY(Vc::abs(ulpDifference - expectedDifference) <= maxUncertainty)
                 << ", base = " << base << ", epsilon = " << eps << ", diff = " << diff;
-            for (int k = 0; k < V::Size; ++k) {
+            for (size_t k = 0; k < V::Size; ++k) {
                 VERIFY(std::abs(ulpDifference[k] - expectedDifference[k]) <= maxUncertainty[k]);
             }
         }
